+++ packages/migrator/src/utils/index.ts
@@ -4,5 +4,7 @@
 
 export * from './config-loader';
 export * from './constants';
+export * from './import-mapping';
 export * from './logger';
 export * from './migration-history';
+export * from './repo-config';